refactor(frontend): decompose and decouple productivity screen
Debt Identified: The `productivity.tsx` screen was a "God Widget" containing 858 lines of code. It violated SRP by handling complex data filtering, sorting, prioritization, and UI rendering in a single file. Several inline "Mega Widgets" caused logic leaks and readability issues.

Refactored Code:
- Decomposed `productivity.tsx` into 5 discrete UI components (`ProductivityHeader`, `PriorityChips`, `EventCard`, `TodayPlanCard`, `ProductivityEmptyState`).
- Decoupled all business logic (filtering notes/tasks/events, calculating task priorities, generating daily plans) into a dedicated `useProductivityViewModel` hook.
- Ensured strong typing and replaced Magic Numbers with design tokens.

Structural Note: This change improves long-term maintainability by strictly enforcing the View-Model pattern and SOLID SRP. The UI now purely maps state to components, while the view-model hook handles data orchestration, making it highly testable and extensible for future productivity features.
